->Implementation of a system to create a language toolbar based on the translations available within the system when merging all the HTML files; ->The grunt tasks are now organized on files; ->Link to the project repository was updated in all the files;
