+++ README.md
@@ -1,6 +1,6 @@
 Presentation about accessible websites development
 ---------
 
-Presentation with an [Introduction to the development of more accessible websites](http://cesperanc.github.io/accessible-websites-presentation/) - in Portuguese.
+Presentation with an [Introduction to the development of more accessible websites](https://cesperanc.github.io/daws/).
