+++ grunt_tasks/concurrent.js
@@ -1,0 +1,10 @@
+module.exports = function (grunt) {
+    grunt.config('concurrent', {
+        css: ['compass:build', 'autoprefixer:build', 'cssmin:build', 'concat:build'],
+        js: ['jshint:build', 'uglify:build'],
+        options: {
+            limit: 5,
+            logConcurrentOutput: true
+        }
+    });
+};
