+++ grunt_tasks/cssmin.js
@@ -1,0 +1,9 @@
+module.exports = function (grunt) {
+    grunt.config('cssmin', {
+        build: {
+            expand: true,
+            src: ['**/*.dev.css'],
+            ext: '.css'
+        }
+    });
+};
