+++ grunt_tasks/uglify.js
@@ -1,0 +1,16 @@
+module.exports = function (grunt) {
+    grunt.config('uglify', {
+        options: {
+            banner: '<%= banner %>'
+        },
+        build: {
+            expand: true,
+            src: ['**/*.dev.js', '!*.min.js'],
+            ext: '.min.js'
+        },
+        from_watch: {
+            src: '*.dev.js',
+            dest: '*.js'
+        }
+    });
+};
