+++ html/en/slides/000-cover/title.html
@@ -3,5 +3,5 @@
     <ul class="authors">
         <li><span class="name">Cl&aacute;udio Esperan&ccedil;a &lt;<a href="mailto:cesperanc@gmail.com?subject=Introduction%20to%20the%development%20of%20accessible%20web%20sites" target="_blank" title="Send an email to Cláudio Esperança">cesperanc@gmail.com</a>&gt;</span></li>
     </ul>
-    <a class="link" href="http://cesperanc.github.io/accessible-websites-presentation/">http://cesperanc.github.io/accessible-websites-presentation/</a>
+    <a class="link" href="https://cesperanc.github.io/daws/">cesperanc.github.io/daws/</a>
 </article>
