+++ html/pt-PT/layout/footer.html
@@ -1,4 +1,5 @@
         </section>
+        <div id="impress-toolbar" class="no-print"><!--[lang-toolbar]--></div>
         <script type="text/javascript">
         //<![CDATA[
             var ImpressL10n = {
