+++ html/pt-PT/slides/000-cover/title.html
@@ -3,5 +3,5 @@
     <ul class="authors">
         <li><span class="name">Cláudio Esperança &lt;<a href="mailto:cesperanc@gmail.com?subject=Introdução%20ao%20desenvolvimento%20de%20sites%20acessíveis" target="_blank" title="Siga a hiperligação para enviar uma mensagem de correio eletrónico a Cláudio Esperança">cesperanc@gmail.com</a>&gt;</span></li>
     </ul>
-    <a class="link" href="http://cesperanc.github.io/accessible-websites-presentation/">http://cesperanc.github.io/accessible-websites-presentation/</a>
+    <a class="link" href="https://cesperanc.github.io/daws/">cesperanc.github.io/daws/</a>
 </article>
