+++ html/pt-PT/slides/999-meta-information/999-source-code.html
@@ -1,6 +1,6 @@
 <article id="source-code" class="step slide" data-x="+0" data-y="1000" data-z="3000" data-rotate-x="0" data-rotate-y="45">
     <h2 class="title"><span class="label">Source code</span></h2>
     <ul>
-        <li>GitHub: <a href="https://github.com/cesperanc/accessible-websites-presentation" target="_blank" title="Ligação com o código-fonte da apresentação">https://github.com/cesperanc/accessible-websites-presentation</a></li>
+        <li>GitHub: <a href="https://github.com/cesperanc/daws" target="_blank" title="Ligação com o código-fonte da apresentação">https://github.com/cesperanc/daws</a></li>
     </ul>
 </article>
