+++ index.html
@@ -37,9 +37,10 @@
     <ul class="authors">
         <li><span class="name">Cl&aacute;udio Esperan&ccedil;a &lt;<a href="mailto:cesperanc@gmail.com?subject=Introduction%20to%20the%development%20of%20accessible%20web%20sites" target="_blank" title="Send an email to Cláudio Esperança">cesperanc@gmail.com</a>&gt;</span></li>
     </ul>
-    <a class="link" href="http://cesperanc.github.io/accessible-websites-presentation/">http://cesperanc.github.io/accessible-websites-presentation/</a>
+    <a class="link" href="https://cesperanc.github.io/daws/">cesperanc.github.io/daws/</a>
 </article>
         </section>
+        <div id="impress-toolbar" class="no-print"><a class="impress-tool lang-selector current-lang" hreflang="en" href="index.html" lang="en">English</a><a class="impress-tool lang-selector" hreflang="pt-PT" href="index_pt-PT.html" lang="pt-PT">Português</a></div>
         <script type="text/javascript">
         //<![CDATA[
             var ImpressL10n = {
