+++ index_pt-PT.html
@@ -37,7 +37,7 @@
     <ul class="authors">
         <li><span class="name">Cláudio Esperança &lt;<a href="mailto:cesperanc@gmail.com?subject=Introdução%20ao%20desenvolvimento%20de%20sites%20acessíveis" target="_blank" title="Siga a hiperligação para enviar uma mensagem de correio eletrónico a Cláudio Esperança">cesperanc@gmail.com</a>&gt;</span></li>
     </ul>
-    <a class="link" href="http://cesperanc.github.io/accessible-websites-presentation/">http://cesperanc.github.io/accessible-websites-presentation/</a>
+    <a class="link" href="https://cesperanc.github.io/daws/">cesperanc.github.io/daws/</a>
 </article>
 <article id="index" class="step slide" data-x="2000" data-y="0" data-z="0">
     <h2 class="title"><span class="label">Índice</span></h2>
@@ -718,10 +718,11 @@ a:active{
 <article id="source-code" class="step slide" data-x="+0" data-y="1000" data-z="3000" data-rotate-x="0" data-rotate-y="45">
     <h2 class="title"><span class="label">Source code</span></h2>
     <ul>
-        <li>GitHub: <a href="https://github.com/cesperanc/accessible-websites-presentation" target="_blank" title="Ligação com o código-fonte da apresentação">https://github.com/cesperanc/accessible-websites-presentation</a></li>
+        <li>GitHub: <a href="https://github.com/cesperanc/daws" target="_blank" title="Ligação com o código-fonte da apresentação">https://github.com/cesperanc/daws</a></li>
     </ul>
 </article>
         </section>
+        <div id="impress-toolbar" class="no-print"><a class="impress-tool lang-selector" hreflang="en" href="index.html" lang="en">English</a><a class="impress-tool lang-selector current-lang" hreflang="pt-PT" href="index_pt-PT.html" lang="pt-PT">Português</a></div>
         <script type="text/javascript">
         //<![CDATA[
             var ImpressL10n = {
