+++ package.json
@@ -13,15 +13,16 @@
     "grunt-contrib-uglify": "^0.9.1",
     "grunt-contrib-watch": "~0.5.3",
     "grunt-csso": "~0.6.3",
-    "grunt-modernizr": "^0.6.0"
+    "grunt-modernizr": "^0.6.0",
+    "load-grunt-tasks": "^3.2.0"
   },
   "name": "accessible-website-development",
   "version": "0.6.0",
   "description": "Presentation with some information about accessible website development",
-  "homepage": "http://cesperanc.github.io/accessible-websites-presentation/",
+  "homepage": "https://cesperanc.github.io/daws/",
   "repository": {
     "type": "git",
-    "url": "git@github.com:cesperanc/accessible-websites-presentation.git"
+    "url": "git@github.com:cesperanc/daws.git"
   },
   "keywords": [
     "Accessibility",
